MAR-1660 added server logic on server for s3 life time link
