+++ server/controllers/S3Controller.js
@@ -1,0 +1,10 @@
+const { getLinkWithLifeTime } = require('../services/S3Service')
+
+async function s3GetLinkWithLifeTime(req, res) {
+  const url = await getLinkWithLifeTime(req.body)
+  return res.json(url)
+}
+
+module.exports = {
+  s3GetLinkWithLifeTime,
+}
