+++ server/routes/api.js
@@ -14,12 +14,14 @@ const storage = multer.diskStorage({
 
 // controllers
 const { create: createLead } = require('../controllers/LeadsController')
+const { s3GetLinkWithLifeTime } = require('../controllers/S3Controller')
 const { index: sendVacancyApplication } = require('../controllers/CareersController')
 const { send } = require('../controllers/SendEmailController')
 
 const router = express.Router()
 
 router.post('/leads', createLead)
+router.post('/get-link-with-life-time', s3GetLinkWithLifeTime)
 router.post('/careers', multer({ storage }).single('cvFile'), sendVacancyApplication)
 router.post('/send-email', multer({ storage, limits: { fieldSize: 25 * 1024 * 1024 } }).single('base64'), send)
